Fix archiver to extractor naming
