+++ src/auto_archiver/modules/instagram_tbot_extractor/__init__.py
@@ -1,1 +1,1 @@
-from .instagram_tbot_archiver import InstagramTbotExtractor
+from .instagram_tbot_extractor import InstagramTbotExtractor
